Create abc.cpp
AC

This problem can be seen as a complete search problem; you just need to narrow down what you are searching.

In this case, you should search the pairwise differences among the given numbers because A, B, and C all MUST be a difference between two of the given numbers.

This observation is actually relatively straightforward if you think about it.