Update diamond.cpp
Reduce O(N^2) to O(N) by iterating only for one of the diamond cases.

Takeaway is to try to have only one thing "moving around" and the others monotonously iterating. Also, use pre-computation (in this case of the max size of the other diamond case and in other cases maybe prefix sums).