Update gifts.cpp
AC

Big idea: needed to find an effective way to draw out the problem.

Needed to check if two nodes could be in a simple cycle (without actually finding the simple cycles).

Initial idea: test each possible gift to be received and recurse, which would be O(N!); equivalent to trying to find all simple cycles in the graph while doing a lot of extra work.